requirements.txt
Pillow
cryptography
zlib